Store the job input files [skip hazardlib][demos]
