+++ openquake/calculators/export/__init__.py
@@ -15,7 +15,8 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
-
+import zipfile
+import io
 from openquake.baselib.general import import_all, CallableDict
 from openquake.commonlib.writers import write_csv
 
@@ -59,8 +60,20 @@ def keyfunc(ekey):
     fullname, ext = ekey
     return (fullname.split('/', 1)[0], ext)
 
+
 export = CallableDict(keyfunc)
 
 export.from_db = False  # overridden when exporting from db
 
 import_all('openquake.calculators.export')
+
+
+@export.add(('job_zip', 'zip'))
+def export_job_zip(ekey, dstore):
+    """
+    Export the bytes in the job_zip dataset in a job.zip file
+    """
+    dest = dstore.export_path('job.zip')
+    zbytes = dstore['job_zip'].value + b'\x00' * 4
+    open(dest, 'wb').write(zbytes)
+    return [dest]
