+++ openquake/commonlib/readinput.py
@@ -116,7 +116,9 @@ def get_params(job_inis, **kw):
     :returns:
         A dictionary of parameters
     """
+    job_zip = None
     if len(job_inis) == 1 and job_inis[0].endswith('.zip'):
+        job_zip = job_inis[0]
         job_inis = extract_from_zip(
             job_inis[0], ['job_hazard.ini', 'job_haz.ini',
                           'job.ini', 'job_risk.ini'])
@@ -132,6 +134,8 @@ def get_params(job_inis, **kw):
     job_ini = os.path.abspath(job_inis[0])
     base_path = decode(os.path.dirname(job_ini))
     params = dict(base_path=base_path, inputs={'job_ini': job_ini})
+    if job_zip:
+        params['inputs']['job_zip'] = os.path.abspath(job_zip)
 
     for sect in cp.sections():
         _update(params, cp.items(sect), base_path)
